refactor test-utils out of class-decor.spec
